#95: changed database connections
